+++ Version 1/Terminal3/Terminal3/ServiceLayer/Controllers/RegisteredUserController.cs
@@ -29,7 +29,7 @@ namespace Terminal3.ServiceLayer.Controllers
         public Result<Boolean> UpdateShoppingCart(String userID, String storeID, String productID, int quantity) { return StoresAndManagementInterface.UpdateShoppingCart(userID,  storeID,  productID,  quantity); }
         public Result<Object> Purchase(String userID, IDictionary<String, Object> paymentDetails, IDictionary<String, Object> deliveryDetails) { return StoresAndManagementInterface.Purchase(userID, paymentDetails, deliveryDetails); }
         public Result<HistoryDAL> GetUserPurchaseHistory(String userID) { return StoresAndManagementInterface.GetUserPurchaseHistory(userID); }
-        public Result<int> GetTotalShoppingCartPrice(String userID) { return StoresAndManagementInterface.GetTotalShoppingCartPrice(userID); }
+        public Result<double> GetTotalShoppingCartPrice(String userID) { return StoresAndManagementInterface.GetTotalShoppingCartPrice(userID); }
         public Result<StoreDAL> OpenNewStore(String storeName, String userID) { return StoresAndManagementInterface.OpenNewStore(storeName, userID); }
 
         #endregion
